order

+++ application/controllers/order.php
@@ -14,7 +14,7 @@ class Order extends Base {
    public function lihatKeranjang(){
       $data = array(
          'title'=>'Lihat Keranjang',
-      );
+         );
       $this->displayUser('order/lihatKeranjang',$data);
    }
    //lihat status order
@@ -26,14 +26,14 @@ class Order extends Base {
                'script'=>'<script>$(document).ready(function(){$("#menunggu").addClass("active");});</script>',
                'title'=>'Status Order',
                'view'=>$this->m_order->myCustomOrder('menunggu pembayaran'),
-            );
+               );
             break;
             case 'lunas':
             $data = array(
                'script'=>'<script>$(document).ready(function(){$("#lunas").addClass("active");});</script>',
                'title'=>'Status Order',
                'view'=>$this->m_order->myCustomOrder('lunas'),
-            );
+               );
             break;
          }
       }else{
@@ -41,7 +41,7 @@ class Order extends Base {
             'script'=>'<script>$(document).ready(function(){$("#semua").addClass("active");});</script>',
             'title'=>'Status Order',
             'view'=>$this->m_order->myOrder(),
-         );
+            );
       }
       $this->displayUser('order/lihatOrder',$data);
    }
@@ -51,7 +51,7 @@ class Order extends Base {
       $data = array(
          'title'=>'LIhat Order Item',
          'view'=>$this->m_order->getOrderItem($idOrder),
-      );
+         );
       $this->displayUser('order/lihatItem',$data);
    }
    ///////////////////////////
@@ -59,14 +59,6 @@ class Order extends Base {
    ///////////////////////////
    //add to cart
    public function addToCart(){
-      if($this->cart->total_items()==0){
-         echo '
-         <script>
-         alert("Cart Masih Kosong");
-            window.location="'.$this->agent->referrer().'";
-         </script>
-         ';
-      }
       $link = explode('?',$this->agent->referrer());
       $link = $link[0];
       if(empty($_POST)){
@@ -83,7 +75,7 @@ class Order extends Base {
                'qty'=>$jumlah,
                'price'=>$sarung['harga'],
                'name'=>$sarung['nama']
-            );
+               );
             $this->cart->insert($data);//insert to cart session
             redirect($link.'?success=berhasil dimasukan ke cart');//balik ke halaman sebelumnya
          }else{//pesanan lebih besar daripada stok
@@ -105,43 +97,53 @@ class Order extends Base {
    //ALL ABOUT ORDER
    //////////////////////////////////////////
    public function addOrder(){
+      $cart =$this->cart->total_items(); 
+      if($cart<=0||empty($cart)){
+         echo '
+         <script>
+            alert("Cart Masih Kosong");
+            window.location="'.$this->agent->referrer().'";
+         </script>
+         ';
+      } else {
       //create new order
-      $data = array(
-         'id_pelanggan'=>$this->session->userdata['userlogin'][0]['id_pelanggan'],
-         'id_admin'=>null,
-         'tanggalOrder'=>date('d-m-y H:i:s'),
-         'tanggalLunas'=>null,
-         'harga'=>$this->cart->total(),
-         'status'=>'menunggu pembayaran'
-      );
+         $data = array(
+            'id_pelanggan'=>$this->session->userdata['userlogin'][0]['id_pelanggan'],
+            'id_admin'=>null,
+            'tanggalOrder'=>date('d-m-y H:i:s'),
+            'tanggalLunas'=>null,
+            'harga'=>$this->cart->total(),
+            'status'=>'menunggu pembayaran'
+            );
       //ada to order
-      $this->db->insert('pesan',$data);
+         $this->db->insert('pesan',$data);
       //get lattest id order by user
-      $idOrder = $this->m_order->getLastIdOrder($this->session->userdata['userlogin'][0]['id_pelanggan']);
+         $idOrder = $this->m_order->getLastIdOrder($this->session->userdata['userlogin'][0]['id_pelanggan']);
       //masukan order item
-      foreach($this->cart->contents() as $i):
-         $item = array(
-            'id_pesan' => $idOrder,
-            'id_sarung' => $i['id'],
-            'jumlah' => $i['qty'],
-            'subtotal' => $i['subtotal']
-         );
+         foreach($this->cart->contents() as $i):
+            $item = array(
+               'id_pesan' => $idOrder,
+               'id_sarung' => $i['id'],
+               'jumlah' => $i['qty'],
+               'subtotal' => $i['subtotal']
+               );
          //memasukan ke order item
          $this->db->insert('pesan_item',$item);
          //mengurangi jumlah sarung yang ditampilkan
          $this->m_order->getStock($i['id'],$i['qty']);
          //clear cart
          $this->cart->destroy();
-      endforeach;
+         endforeach;
       redirect(site_url('order/lihatOrder').'?success=order sudah dimasukan, silahkan melakukan pembayaran untuk lanjut ke langkah berikutnya');//balik ke halaman sebelumnya
    }
+}
 
    /////////////////////////////
    // ALL ABOUT CETAK
    /////////////////////////////
-   public function cetakbukti(){
-      $idOrder = $this->uri->segment(3);
-      $data = array(
+public function cetakbukti(){
+   $idOrder = $this->uri->segment(3);
+   $data = array(
          'order'=>$this->m_order->detailOrder($idOrder),//detail order
          'item'=>$this->m_order->getOrderItem($idOrder),//item order
          );
@@ -150,8 +152,8 @@ class Order extends Base {
       $html = $this->output->get_output();
       $this->load->library('dompdf_gen');
 		// Convert to PDF
-		$this->dompdf->load_html($html);
-		$this->dompdf->render();
+      $this->dompdf->load_html($html);
+      $this->dompdf->render();
 		$title = 'Cetak Bukti Order ';//pdf title
 		$this->dompdf->stream($title.".pdf");//pdf file name
    }
